refactoring

+++ src/main/java/com/mitra/controller/impl/post/DeleteTrackController.java
@@ -2,12 +2,16 @@ package com.mitra.controller.impl.post;
 
 import com.mitra.controller.GetUrl;
 import com.mitra.controller.impl.PostController;
+import com.mitra.controller.impl.util.SessionAttrAccessor;
+import com.mitra.dto.TrackDto;
 import com.mitra.service.TrackService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
+import java.util.Optional;
 
 public class DeleteTrackController implements PostController {
 
@@ -19,10 +23,17 @@ public class DeleteTrackController implements PostController {
 
     @Override
     public void processPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // todo : check user is owner of song
-
+        int myId = SessionAttrAccessor.getProfileId(request);
         int trackId = Integer.parseInt(request.getParameter("trackId"));
-        trackService.remove(trackId);
+        List<TrackDto> profileMusic = trackService.getProfileMusic(myId);
+
+        Optional<TrackDto> optionalTrack = profileMusic.stream()
+                .filter(track -> track.getId() == trackId)
+                .findFirst();
+
+        if (optionalTrack.isPresent()) {
+            trackService.remove(trackId);
+        }
 
         response.sendRedirect(GetUrl.MUSIC.getUrl());
     }
